Refactor all classes into their packages
