+++ src/main/java/aisu/parser/Parser.java
@@ -1,5 +1,6 @@
-package aisu;
+package aisu.parser;
 
+import aisu.tasklist.TaskList;
 import aisu.command.AddCommand;
 import aisu.command.Command;
 import aisu.command.DeleteCommand;
@@ -7,6 +8,7 @@ import aisu.command.ExitCommand;
 import aisu.command.FindCommand;
 import aisu.command.MarkDoneCommand;
 import aisu.command.ShowListCommand;
+import aisu.exception.AisuException;
 
 /**
  * A Parser for text inputs.
