+++ src/main/java/aisu/storage/Storage.java
@@ -1,5 +1,7 @@
-package aisu;
+package aisu.storage;
 
+import aisu.tasklist.TaskList;
+import aisu.exception.AisuException;
 import aisu.task.Deadline;
 import aisu.task.Event;
 import aisu.task.Task;
