+++ src/main/java/aisu/ui/gui/MainApp.java
@@ -1,4 +1,4 @@
-package aisu.gui;
+package aisu.ui.gui;
 
 import java.io.IOException;
