+++ src/main/resources/view/MainWindow.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1" fx:controller="aisu.gui.MainWindow">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1" fx:controller="aisu.ui.gui.MainWindow">
   <children>
     <TextField fx:id="userInput" layoutX="8.0" layoutY="557.0" onAction="#handleUserInput" prefHeight="28.0" prefWidth="307.0" AnchorPane.bottomAnchor="9.0" />
     <Button fx:id="sendButton" layoutX="324.0" layoutY="561.0" mnemonicParsing="false" onAction="#handleUserInput" prefHeight="31.0" prefWidth="69.0" text="Send" />
