+++ src/test/java/aisu/StorageTest.java
@@ -1,5 +1,6 @@
 package aisu;
 
+import aisu.exception.AisuException;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
